More Bits
Bits now have appear as either 0 or 1 and are assigned to the octree
properly. This required two separate particle systems which are handled
together in the GristManager. Future work: Combine the two bit arrays
when looping through them.
Bit grist steers towards the player and deallocates itself from the
octree when it passes through the player (using a raycast between
previous and new positions on the player's layer).

Allocation replaced with reallocation, as particles automatically get
reassigned in Unity ParticleSystems so the indexes must also be
reallocated in the octree.

Made the camera 3rd person rather than 1st so that grist collection
doesn't swamp the screen with particles.