Player Movement and Tube Model
Player movement made more like Strike Suit Zero.
Simulated cursor is constrained to ELLIPSE (not a circle, unless your
aspect ratio is 1:1) and ship appears to aim towards the cursor.
Ship angularly accelerates based on the distance of the cursor from the
centre of the screen. Ship rotation is constrained so that collisions
don't rotate it and muck with the player's sense of direction.
Rotation is instead done by setting the rotation each frame.
Movement is done with acceleration applied to the ship's rigidbody.
Tube model made and debug tube prefab replaced with tube model +
corresponding mesh colliders.
Camera ensured to never clip anything by sphere-casting from the ship
position to the camera position with a sphere the size of the camera's
near-clip distance and pushing the camera one near-clip-sphere away from
the hit point.
All tubes generated in children of current room rather than just the
tube to the child from the current room.
These tubes have their length initialised to the furthest distance rooms
can be from each other, and the length is changed when the room is
entered.