Movement Change + Room Shader + U-Turn Bounds
Player movement made more swanky and imho more fun. Handles more like a
sci fi spaceship.

Room walls that aren't occupied by tubes have a shader that renders a
skybox infinitely far away and a grid in world space that fades in when
the camera approaches it.

Instead of colliding with the unoccupied room walls solidly, the ship
now does a U-Turn to get back while the camera sits back to make the
user aware that they can't go that way. I've forgotten the name of the
game that does this, but I'm sure an old game did it.